Add requirements.txt, /example and updated main.py
Main.py now has a function, still very crude, that reads text from a pdf.
For testing purposes the pdf location is hardcoded but I do not intend
on keeping it that way.

example directory stores a copy of Alice in Wonderland that I can use for
testing and as an example for any users.

requirements.txt will contain any required python libraries